+++ liquity/settings.py
@@ -173,3 +173,14 @@ AUTH_USER_MODEL = 'users.User'
 # Login
 LOGIN_REDIRECT_URL = '/drinks'
 LOGIN_URL = '/login/'
+
+
+# Email
+EMAIL_USE_TLS = True
+EMAIL_HOST = 'smtp.gmail.com'
+EMAIL_HOST_USER = config["EMAIL_USER"]
+EMAIL_HOST_PASSWORD = config["EMAIL_PASSWORD"]
+EMAIL_PORT = 587
+
+# For Debugging
+EMAIL_BACKEND = 'django.core.mail.backends.console.EmailBackend'