+++ liquity/urls.py
@@ -32,7 +32,13 @@ urlpatterns = i18n_patterns(
     path('logout/', auth_views.LogoutView.as_view(template_name="users/logout.html"), name="logout"),
 
     path('password/change/', user_views.password_change, name='change_password'),
+    path('password/reset/', auth_views.PasswordResetView.as_view(template_name='users/password_reset.html'), name='password_reset'),
+    path('password/reset/done/', auth_views.PasswordResetDoneView.as_view(template_name='users/password_reset_done.html'), name='password_reset_done'),
+    path('password/reset/<uidb64>/<token>/', auth_views.PasswordResetConfirmView.as_view(template_name='users/password_reset_confirm.html'), name='password_reset_confirm'),
+    path('password/reset/complete/', auth_views.PasswordResetCompleteView.as_view(template_name='users/password_reset_complete.html'), name='password_reset_complete'),
+
     path('register/', user_views.register, name="register"),
+
     path('profile/<uuid:pk>/', user_views.profile, name="profile"),
     path('friends/', user_views.friends, name="friends"),
     path('search/', user_views.search, name="search"),
@@ -40,7 +46,7 @@ urlpatterns = i18n_patterns(
     path('', include('app.urls')),
     path('api/', include('api.urls')),
     path('actions/', include('actions.urls')),
-    path('', include('django.contrib.auth.urls')),
+    # path('', include('django.contrib.auth.urls')),
 )
 
 if settings.DEBUG: